Question_form & Search Feature & CRUD operations restricted

+++ base/views.py
@@ -1,13 +1,23 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from .models import Room, Question
-from .forms import RoomForm
-
+from .forms import RoomForm, QuestionForm
+from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''  # for searching
+    # rooms = Room.objects.filter(name__icontains=q)
+    rooms = Room.objects.filter(
+        Q(name__icontains=q)  #|
+        # Q(description__icontains=q) #add search by user too
+    )
+
+    room_count=rooms.count()
+
+    context = {'rooms': rooms,'room_count':room_count}
     return render(request, "home.html", context)
 
 
@@ -24,22 +34,26 @@ def question(request, pk, pk2):
     context = {'question': question}
     return render(request, "question.html", context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
         form = RoomForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect('home')
+            return redirect('createQuestion')
 
     context = {'form': form}
     return render(request, "form.html", context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    # checking for valid host
+    if request.user != room.host:
+        return HttpResponse("You can't do this ...")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -50,11 +64,28 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, "form.html", context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    # checking for valid host
+    if request.user != room.host:
+        return HttpResponse("You can't do this ...")
 
     if request.method == "POST":
         room.delete()
         return redirect('home')
 
-    return render(request, "delete.html", {'obj':room})
+    return render(request, "delete.html", {'obj': room})
+
+@login_required(login_url='login')
+def createQuestion(request):
+    form = QuestionForm()
+    if request.method == "POST":
+        form = QuestionForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('createQuestion')
+
+    context = {'form': form}
+    return render(request, "question_form.html", context)
